% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_and_pre-processing.R
\name{check_input_geNet}
\alias{check_input_geNet}
\title{checking input df Function}
\usage{
check_input_geNet(binary_df)
}
\arguments{
\item{binary_df}{a binary dataframe to be checked. Mandatory argument.}
}
\value{
None
}
\description{
function to check the format of the input binary df. It is automatically called by the geNet() function.
}
\examples{
\dontrun{check_input_df(binary_df)}
}
