% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking_and_pre-processing.R
\name{check_input_visualization}
\alias{check_input_visualization}
\title{check_input_visualization}
\usage{
check_input_visualization(data)
}
\arguments{
\item{data}{output of the geNet() function (list of two ffdf objects)}
}
\value{
error message if checking fails
}
\description{
function to check the format of the input binary df. It is automatically called by the geNet() function.
}
\examples{
\dontrun{check_input_visualization(data)}
}
