% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visnetwork.R
\name{contract_network}
\alias{contract_network}
\title{contract_network}
\usage{
contract_network(data, seed = 123, show_label = F, show_legend = T)
}
\arguments{
\item{data}{visnetwork dataframe generated by the geNet algorithm. Mandatory argument.}

\item{seed}{set seed of the nodes}

\item{show_label}{should the labels of the nodes be visible? default to False}

\item{show_legend}{should the legend be visible? default to True}
}
\value{
list of two objects:
\itemize{
\item visnet: object of class "visNetwork" generated from the contracted data.
\item data_contracted: a compressed version of the input data object. It is generated by contracting the nodes of each cluster in one vertex.
}

This dataset reports only the nodes contained in each cluster and the connections between the clusters
}
\description{
function to contract the visnetwork dataframe. It is automatically called by the plot_visnetwork() function if contract_net=T.
}
