% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNet_algorithm.R
\name{export_geNet_result}
\alias{export_geNet_result}
\title{Export geNet results}
\usage{
export_geNet_result(geNet_output, out_directory, format = "csv")
}
\arguments{
\item{geNet_output}{the object generated by the geNet function. It is a list of two ffdf objects.
\itemize{
\item nodes: ffdf object containing the information about the nodes
\item edges: ffdf object containing the information about the edges
}}

\item{out_directory}{the directory to write the output}

\item{format}{the user can export the geNet output in .csv format or in .ffdata format.
In both cases, two files are generated. One containing the nodes information (e.g., clustering association),
and the other one containing edges information (e.g.,connections weight).
Default to "csv".}
}
\value{
In case of the csv format, two csv files are generated in the output directory
\itemize{
\item nodes_geNet_output.csv: it contains the information about the nodes
\item edges_geNet_output.csv: it contains the information about the edges
}
}
\description{
function to export the geNet output.
It is automatically called by the geNet() function if export_results=T
}
\examples{
\dontrun{export_geNet_result(geNet_output,out_directory="./",format="csv") 
# export output in current working directory "./"
}
}
