% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_score_df.R
\name{final_score}
\alias{final_score}
\title{function to format the final df of scores}
\usage{
final_score(
  final_df_phi_score,
  sel_weight = "coeff",
  pval_thr_pos = 0.01,
  pval_thr_neg = 0.1
)
}
\arguments{
\item{final_df_phi_score}{the dataframe of scores generated by the generate_final_df() function. Mandatory argument}

\item{sel_weight}{select the type of weights of the edges. Possible values:
logpvalue: choose the negative log-pvalues as weights of the edges
coeff: choose the correlation coefficients as weights of the edges
Default to "coeff"}

\item{pval_thr_pos}{threhold p-value positive edges. Default to 0.01.}

\item{pval_thr_neg}{threshold p-value negative edges. Default to 0.1.}
}
\value{
Object of class "ffdf", containing the connections between the nodes.
The connections have been filtered and correctly formatted to be converted in an igraph object
}
\description{
function to format the final df of scores. The adjustment and filtering steps are executed.
It is automatically called by the geNet() function
}
\examples{
\dontrun{final_score(final_df_phi_score,sel_weight="coeff")}
}
