% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_igraph_network.R
\name{gen_network_obj}
\alias{gen_network_obj}
\title{generate igraph network}
\usage{
gen_network_obj(final_df_score)
}
\arguments{
\item{final_df_score}{the dataframe of scores generated by the final_score() function. Mandatory argument.}
}
\value{
Object of class "igraph"
}
\description{
function to generate the igraph network object based on the final scores.
It is called automatically by the geNet() function.
}
