% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visnetwork_conv.R
\name{gen_visnetwork_data}
\alias{gen_visnetwork_data}
\title{gen_visnetwork_data}
\usage{
gen_visnetwork_data(igraph_network)
}
\arguments{
\item{igraph_network}{igraph object. Mandatory argument.}
}
\value{
list of two objects of class "ffdf".
\itemize{
\item nodes: reports info about the nodes and clustering
\item edges: reports info about the connections between the nodes
This format is designed to be used with the visNetwork package.
}
}
\description{
function to get the object to generate the visnetwork. It is automatically called by the geNet() function.
}
