% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_weights_edges.R
\name{generate_df_scores}
\alias{generate_df_scores}
\title{generate_df_scores}
\usage{
generate_df_scores(all_combs, binary_matrix, signif_test = "cor_test")
}
\arguments{
\item{all_combs}{matrix of combinations}

\item{binary_matrix}{binary matrix of presence/absences genes}

\item{signif_test}{significance test to use.
\itemize{
\item cor_test: use Pearson correlation test approach (faster)
\item chisquare: use chi-square test approach (slower)
Read vignettes for additional details about the significance tests used.
}}
}
\value{
Object of class "dataframe" containing the connections between the nodes and relative weights
}
\description{
function to generate the weights of the edges given the combinations.
It is automatically called  by the geNet() function.
the correlation coefficient and its pvalue is calculated using the cor.test function from the stat package.
See the relative documentation for additional details.
}
