% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_weights_edges.R
\name{generate_final_df}
\alias{generate_final_df}
\title{generate_final_df

function to generate the final df of scores based on the weights of edges
generated by the generate_df_scores_ff() function.
It is called automatically by the geNet() function}
\usage{
generate_final_df(binary_matrix, n_cores = 1, test = "cor_test")
}
\arguments{
\item{binary_matrix}{binary matrix of presence/absences genes}

\item{n_cores}{number of cores to use. Default to 1.}

\item{test}{significance test to use.
\itemize{
\item cor_test: use Pearson correlation test approach (faster)
\item chisquare: use chi-square test approach (slower)
}}
}
\value{
Object of class "ffdf" containing the connections between the nodes and relative weights
}
\description{
generate_final_df

function to generate the final df of scores based on the weights of edges
generated by the generate_df_scores_ff() function.
It is called automatically by the geNet() function
}
\examples{
\dontrun{generate_final_df(binary_matrix,progress=F,n_cores=1,test="cor_test")}
}
