% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_genes.R
\name{generate_freqs_genes}
\alias{generate_freqs_genes}
\title{Calculate genes frequency}
\usage{
generate_freqs_genes(input_binary_df)
}
\arguments{
\item{input_binary_df}{Object of class "dataframe", containing the absence/presence of genes}
}
\value{
Object of class "dataframe". The first column reports the ID of the nodes/genes,
the second column reports the frequency across all strains
}
\description{
function to calculate the percentage of occurrence of every gene across
all strains.
}
\examples{
\dontrun{generate_freqs_genes(input_binary_df)}
}
