% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers_values.R
\name{get_colors_based_on_size}
\alias{get_colors_based_on_size}
\title{get_colors_based_on_size}
\usage{
get_colors_based_on_size(
  data,
  size_thr,
  direction = "less",
  col_nodes = "black"
)
}
\arguments{
\item{data}{List of two ffdf objects (format generated by the geNet() function )}

\item{size_thr}{size threshold. Mandatory argument.}

\item{direction}{values greater/less  than (or equal) the selected threshold}

\item{col_nodes}{color of the nodes that match the specified condition. Default to "black".}
}
\value{
data_new_colors: object of class "dataframe", which reports the new colors for each node of the network
}
\description{
function to get the colors based on the size of the nodes
}
\examples{
\dontrun{get_colors_based_on_size(data,size_thr=10,direction="less",col_nodes="black")}
}
