% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers_values.R
\name{get_colors_based_on_strains}
\alias{get_colors_based_on_strains}
\title{get new colors based on the strains}
\usage{
get_colors_based_on_strains(
  input_binary_df,
  data,
  strains_names,
  col_nodes = "black"
)
}
\arguments{
\item{input_binary_df}{Object of class "dataframe", containing presence/absence of genes. Mandatory argument.}

\item{data}{visnetwork dataframe generated by the geNet algorithm. Mandatory argument.}

\item{strains_names}{the vector containing the names of the strains to visualize. Mandatory argument.}

\item{col_nodes}{Color of the nodes matched with the strains. Default to "black".}
}
\value{
Object of class "dataframe", which reports the new colors for each node of the network
}
\description{
function to predict the colors based on the strains occurrence of genes
i.e.,show nodes/genes associated with the selected strains (i.e., that gene is present in that strains)
}
\examples{
\dontrun{get_colors_based_on_strains(input_binary_df,data,strains_names=c("strain_x"),col_nodes="black")}
}
