% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers_values.R
\name{get_gene_ID_based_on_tooltip}
\alias{get_gene_ID_based_on_tooltip}
\title{get_gene_ID_based_on_tooltip}
\usage{
get_gene_ID_based_on_tooltip(data, vec_tooltip)
}
\arguments{
\item{data}{List of two ffdf objects (format generated by the geNet() function )}

\item{vec_tooltip}{vector of annotation to match with the gene IDs. Mandatory argument.}
}
\value{
Object of class "dataframe", which reports the genes ID for the specified annotation keywords.
}
\description{
function to get the gene IDs based on the annotation (tooltip column)
}
\examples{
\dontrun{get_gene_ID_based_on_tooltip(data,vec_tooltip=c("protein","secretion"))
# Get the nodes IDs whose annotation reports the word "protein" and "secretion"
}
}
