% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers_values.R
\name{get_group_names_based_on_n_nodes}
\alias{get_group_names_based_on_n_nodes}
\title{get the group names based on the number of nodes}
\usage{
get_group_names_based_on_n_nodes(data)
}
\arguments{
\item{data}{List of two ffdf objects (format generated by the geNet() function )}
}
\value{
Object of class "dataframe", which reports the old names associated with the new names
}
\description{
function to get the group names based on the number of nodes in each group.
}
\examples{
\dontrun{get_group_names_based_on_n_nodes(data)}
}
