% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers_values.R
\name{get_group_names_based_on_tooltip}
\alias{get_group_names_based_on_tooltip}
\title{get group names based on annotation}
\usage{
get_group_names_based_on_tooltip(data)
}
\arguments{
\item{data}{List of two ffdf objects (format generated by the geNet() function )}
}
\value{
Object of class "dataframe", which reports the old names associated with the new names
}
\description{
function to get the group names based on the annotation of the nodes (tooltip column)
}
\examples{
\dontrun{get_group_names_based_on_tooltip(data)}
}
