% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers_values.R
\name{get_groups_based_on_annotation}
\alias{get_groups_based_on_annotation}
\title{get groups based on the annotation}
\usage{
get_groups_based_on_annotation(data, select_annotation)
}
\arguments{
\item{data}{List of two ffdf objects (format generated by the geNet() function )}

\item{select_annotation}{vector of annotation words to match with the groups}
}
\value{
Object of class "vector", which reports the groups selected
}
\description{
function to get the groups based on the annotation (tooltip column)
}
\examples{
\dontrun{get_groups_based_on_tooltip(data,select_tooltip=c("protein","secretion"))
# Get the groups where the annotations words "protein" and "secretion" appear.
}
}
