% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers_values.R
\name{get_groups_based_on_clustering}
\alias{get_groups_based_on_clustering}
\title{get_groups_based_on_clustering}
\usage{
get_groups_based_on_clustering(igraph_network, method = "infomap")
}
\arguments{
\item{igraph_network}{igraph network generated by the gen_igraph_network() function. Mandatory argument.}

\item{method}{clustering method:
\itemize{
\item louvain: predict clusters using the Louvain algorithm.
The "modularity" is the objective function to maximize. Small clusters may be hidden. It detects only the bigger clusters.
\item infomap: predict clusters using the Infomap algorithm.
The "map equation" is the objective function to minimize. Compared to the Louvain algorithm, it is able to identify smaller clusters (subclusters)
\item fastgreedy: predicts clusters using the Fastgreedy algorithm
\item walktrap: predicts clusters using the Walktrap algorithm
}

Default to "infomap" to detect smaller clusters. The Louvain algorithm is faster than the Infomap algorithm in large networks.}
}
\value{
Object of class "dataframe", which reports the genes ID associated with the new groups
}
\description{
function to get the groups based on a clustering algorithm
Note: the color of the nodes is based on the group (using the function get_new_colors_based_on_groups())
to evaluate the clustering method we use the modularity measure.
The modularity of a graph with respect to some division (or vertex types) measures how good the division is,
or how separated are the different vertex types from each other.
}
\examples{
\dontrun{get_groups_based_on_clustering(igraph_network,method="infomap")}
}
