% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers_values.R
\name{get_groups_based_on_gene_ID}
\alias{get_groups_based_on_gene_ID}
\title{get_groups_based_on_gene_ID}
\usage{
get_groups_based_on_gene_ID(data, vec_gene_ids)
}
\arguments{
\item{data}{List of two ffdf objects (format generated by the geNet() function )}

\item{vec_gene_ids}{vector of genes IDs to which the groups must be matched. Mandatory argument.}
}
\value{
Object of class "dataframe", which reports the groups for the specified genes ID.
}
\description{
function to get the groups based on the ID of the nodes
}
\examples{
\dontrun{get_groups_based_on_gene_ID(data,vec_gene_ids=c("group_200","group_253"))
# Get the groups for the nodes IDs "group_200" and "group_253"
}
}
