% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visnetwork.R
\name{get_legend_df}
\alias{get_legend_df}
\title{get_legend_df}
\usage{
get_legend_df(data)
}
\arguments{
\item{data}{list of two ffdf objects generated by the geNet algorithm. Mandatory argument.}
}
\value{
Object of class "dataframe", which reports the legend information of the nodes
}
\description{
function to get the legend of the groups. It is automatically called by the plot_visnetwork() function.
}
