% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_visnetwork.R
\name{get_mappings}
\alias{get_mappings}
\title{get_mappings}
\usage{
get_mappings(igraph_net, data)
}
\arguments{
\item{igraph_net}{igraph object. Mandatory argument.}

\item{data}{list of two ffdf objects generated by the geNet algorithm. Mandatory argument.}
}
\value{
Object of "vector" containing the mapping codes.
}
\description{
function to get the correct mapping to contract the network.
It is automatically called by the contract_network() function.
}
