% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layers_values.R
\name{get_new_colors_based_on_groups}
\alias{get_new_colors_based_on_groups}
\title{get_new_colors_based_on_groups}
\usage{
get_new_colors_based_on_groups(df_groups)
}
\arguments{
\item{df_groups}{object of class "dataframe", which reports the genes ID associated with the groups. Mandatory argument.}
}
\value{
Object of class "dataframe", which reports the gene IDs and the groups associated with the colors
}
\description{
function to get the colors based on the groups
}
\examples{
\dontrun{get_new_colors_based_on_groups(df_groups)}
}
