% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_layers.R
\name{mod_group_layer}
\alias{mod_group_layer}
\title{mod_group_layer}
\usage{
mod_group_layer(data, data_new_groups = NULL)
}
\arguments{
\item{data}{List of two ffdf objects (format generated by the geNet() function )}

\item{data_new_groups}{dataframe that reports the new group for each node}
}
\value{
data:  data object modified with the new group layer
}
\description{
function to modify the group layer (influence also the structure of the contracted network)
}
\examples{
\dontrun{mod_group_layer(data,data_new_groups)}
}
