% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_layers.R
\name{mod_group_names}
\alias{mod_group_names}
\title{mod_group_names}
\usage{
mod_group_names(data, df_new_groups_names = NULL)
}
\arguments{
\item{data}{List of two ffdf objects (format generated by the geNet() function )}

\item{df_new_groups_names}{dataframe that reports the new names for each old group name}
}
\value{
data:  data object modified with the new group names for the current group layer
}
\description{
function to modify the names of the group
}
\examples{
\dontrun{mod_group_names(data,df_new_groups_names)}
}
