% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_layers.R
\name{mod_size_layer}
\alias{mod_size_layer}
\title{mod_size_layer}
\usage{
mod_size_layer(data, data_new_size = NULL)
}
\arguments{
\item{data}{List of two ffdf objects (format generated by the geNet() function )}

\item{data_new_size}{dataframe that reports the size for each node}
}
\value{
data:  data object modified with the new size layer
}
\description{
function to modify the size layer (size of each node)
}
\examples{
\dontrun{mod_size_layer(data,data_new_size)}
}
