% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_layers.R
\name{mod_tooltip_visnetwork}
\alias{mod_tooltip_visnetwork}
\title{mod_tooltip_visnetwork}
\usage{
mod_tooltip_visnetwork(data, data_new_tooltip = NULL)
}
\arguments{
\item{data}{List of two ffdf objects (format generated by the geNet() function )}

\item{data_new_tooltip}{dataframe that reports the new tooltip string for each node}
}
\value{
data:  data object modified with the new tooltip layer
}
\description{
function to modify the tooltip layer of the network
}
\examples{
\dontrun{mod_tooltip_visnetwork(data,data_new_tooltip)}
}
