% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-law.R
\name{power_law}
\alias{power_law}
\alias{dpower_law}
\alias{rpower_law}
\alias{qpower_law}
\title{The Power Law Distribution}
\usage{
dpower_law(x, alpha, x_min, x_max, step = 1)

rpower_law(n, alpha, x_min, x_max, step = 1)

qpower_law(p, alpha, x_min, x_max, step = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{alpha}{power law's exponent. Must be positive.}

\item{x_min, x_max}{lower and upper bounds of the distribution.
Must be finite.}

\item{step}{number; size of increment of the sequence defining the support.}

\item{n}{number of observations.}

\item{p}{vector of probabilities.}
}
\value{
\code{dpowerlaw} gives the pmf, \code{qpowerlaw} gives the
quantile function, and \code{rpowerlaw} generates random deviates.
}
\description{
These functions provide information about a power law
distribution on \code{seq(x_min, x_max)} with exponent \code{alpha}.
\code{dpowerlaw} gives the pmf, \code{qpowerlaw} gives the
quantile function, and \code{rpowerlaw} generates random deviates.
}
\details{
The pmf of the distribution reads:
\deqn{p(x;\alpha) = x^{-\alpha} / C,}
for \eqn{x = x_min, x_min + 1, \dots, x_max}, \eqn{\alpha>0}, and where
\eqn{C>0} is a normalizing constant.
}
