% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{visualize_network}
\alias{visualize_network}
\title{visualize_network}
\usage{
visualize_network(
  data,
  select_group = NULL,
  select_tooltip = "None",
  select_size = "None",
  clust_method = "None",
  select_ID = NULL
)
}
\arguments{
\item{data}{List of two ffdf objects (format generated by the geNet() function ). Mandatory argument}

\item{select_group}{select the groups to filter data. Optional argument.}

\item{select_tooltip}{select the annotation keywords to filter data. Optional argument.
Note: the tooltip values are the values that appear when hovering the nodes}

\item{select_size}{select the size of the nodes to filter data. Optional argument.}

\item{clust_method}{select the clustering method. Default to Infomap.}

\item{select_ID}{select the IDs to filter data. Optional argument.}
}
\value{
data: data object filtered
}
\description{
function to visualize the network based on the connections and different conditions
}
\examples{
\dontrun{
data_subsetted<-visualize_network(data,select_group=c("group_1","group_5"))
data_subsetted<-visualize_network(data,select_tooltip=c("protein","secretion"))
data_subsetted<-visualize_network(data,clust_method="louvain")
}
}
