% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_balanced_interactions.R
\name{generate_balanced_assignments}
\alias{generate_balanced_assignments}
\title{Generate balanced interactions}
\usage{
generate_balanced_assignments(K, L, K_each = NA, max_ite = 1000)
}
\arguments{
\item{K}{number of communities}

\item{L}{number of types}

\item{K_each}{number of communities each type should be assigned to}

\item{max_ite}{number of maximum attempts to generate assignments}
}
\value{
A L-by-K sparse matrix of type community assignments.
}
\description{
Generate balanced interactions
}
