% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{genet}
\alias{genet}
\title{Generate Networks}
\usage{
genet(
  n,
  assignments,
  cards,
  dc = TRUE,
  target_centrality,
  target_degree = NA,
  verbose = FALSE,
  return_root = FALSE
)
}
\arguments{
\item{n}{number of random networks to generate.}

\item{assignments}{matrix with non-negative entries and whose rows encode the
membership assignments of types.}

\item{cards}{vector of type cardinalities. Must be whole numbers.}

\item{dc}{logical. If \code{TRUE}, generate networks with target degree
centrality; if \code{FALSE}, generate networks with target eigencentrality.}

\item{target_centrality}{vector of target centralities.}

\item{target_degree}{target average expected degree. This argument is
required only when generating networks with target eigencentrality (i.e.
\code{dc = FALSE}).}

\item{verbose}{logical. If \code{TRUE}, give verbose output.}

\item{return_root}{logical. Should the degree correction terms be returned?}
}
\value{
A list of length \code{n} of \code{igraph} graph objects.
}
\description{
Generate networks with overlapping communities and target node centralities.
}
\details{
The rows of the matrix argument `assignments` are normalized so as
to have unit norm.
}
