% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geNet_algorithm.R
\name{geNet}
\alias{geNet}
\title{Execution geNet algorithm}
\usage{
geNet(
  input_binary_df,
  clust_method = "infomap",
  type_weight = "coeff",
  cores = 1,
  out_dir,
  export_results = F,
  test_pvalue = "cor_test",
  pval_thr_pos = 0.01,
  pval_thr_neg = 0.1
)
}
\arguments{
\item{input_binary_df}{input binary dataframe containing genes occurrences. Mandatory argument.
input_binary_df is an object of class "dataframe". The rownames are the strains names, the columns names are unique genes/nodes IDs.}

\item{clust_method}{clustering method. Default to "infomap".}

\item{type_weight}{possible values:
\itemize{
\item logpvalue: the weights of the edges will be the negative log adjusted p-values
\item coeff: the weights of the edges will be the correlation coefficients values
Note: This is true only for the positive edges. The negative edges will have 0 weights in both cases.
The negative edges doesn't influence the topology of the network or the clustering.
}}

\item{cores}{number of cores to use for parallel processing}

\item{out_dir}{if export_results=T, select output directory}

\item{export_results}{exporting geNet output in csv files?}

\item{test_pvalue}{significance test to use.
\itemize{
\item cor_test: use Pearson correlation test approach (faster)
\item chisquare: use chi-square test approach (slower)
}}

\item{pval_thr_pos}{threhold p-value positive edges. Default to 0.01.}

\item{pval_thr_neg}{threshold p-value negative edges. Default to 0.1.}
}
\value{
list of two objects of class "ffdf".
\itemize{
\item nodes: reports info about the nodes and clustering
\item edges: reports info about the connections between the nodes
This format is designed to be used with the visNetwork package.
}
}
\description{
function to execute the geNet algorithm on a binary dataframe of genes occurrences (presence/absence data)
}
\examples{
\dontrun{geNet(input_binary_df,clust_method="infomap",type_weight="coeff",cores=4)}
}
